+++ src/main.cpp
@@ -20,7 +20,23 @@
 #include "game.h"
 #endif
 
+#ifdef _WIN32
+#include <windef.h>
+#include <winbase.h>
+#include <wincon.h>
+#endif
+
 int main(int argc, char **argv) {
+#ifdef _WIN32
+  // This allows us to write to a cmd.exe or powershell if we were run from
+  // one, but otherwise don't open another window.
+  if (AttachConsole(ATTACH_PARENT_PROCESS)) {
+    freopen("CONIN$", "r", stdin);
+    freopen("CONOUT$", "w", stdout);
+    freopen("CONOUT$", "w", stderr);
+  }
+#endif
+
   const char *path;
 #ifndef EXPORT
   if (argv) {